CI compiling error fixed

+++ src/operator/numpy/random/np_uniform_op.h
@@ -183,7 +183,7 @@ inline bool NumpyUniformOpShape(const nnvm::NodeAttrs &attrs,
       oshape_vec.emplace_back(size[i]);
     }
     SHAPE_ASSIGN_CHECK(*out_attrs, 0, TShape(oshape_vec));
-    for (int input_idx = 0; input_idx < in_attrs->size(); input_idx++) {
+    for (size_t input_idx = 0; input_idx < in_attrs->size(); input_idx++) {
       CheckBroadcastable((*in_attrs)[input_idx], (*out_attrs)[0]);
     }
   } else {
